ai-server/requirements.txt
ultralytics==8.2.0
opencv-python-headless==4.9.0.80
supabase==2.4.0
python-dotenv==1.0.1
numpy==1.26.4
requests==2.31.0
